Split test_euler
